requirements.txt
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.11.2